Allows stores with partial results to be displayed.
For example, if there are 3 measures : 1, 10, 50 graphs / store; and one store lacks one or more measure (for example, sleepycat lacks 50 graphs / store).
Then all stores will be displayed, but for sleepy cat only measure for 1 and 10 graphs / store will appear.

An logging warning is raised when trying to get a measure that a store doesn't have (50 graphs / store for sleepycat in above example).